requirements.txt
streamlit
google-generativeai
scipy
pandas